[SofaPython] a new version of module re-import (incl. a test!) Nothing is done in your back, you need to call "SofaPython.unloadModules()" where you want to unload the already imported modules.
